refactor: rename component folder to src
