+++ blockchain/src/registry/registry_service.py
