+++ blockchain/views/nodes.py
@@ -3,7 +3,7 @@ import json, logging
 from rest_framework.views import APIView
 
 
-from ..components.registry.registry_service import RegistryService
+from ..src.registry.registry_service import RegistryService
 
 from blockchain.libraries.peers_manager import PeersManager
 from ..libraries.blockchain import Blockchain
